% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drildowndemo.R
\name{drilldowndemo_open}
\alias{drilldowndemo_open}
\title{Opens an example Shiny app in the RStudio IDE}
\usage{
drilldowndemo_open(app_name = NULL)
}
\arguments{
\item{app_name}{Name of one of Shiny apps within the package. Possible values:
intro, tabs, navigation, details}
}
\description{
Opens an example Shiny app in the RStudio IDE
}
