% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drildowndemo.R
\name{drilldowndemo_run}
\alias{drilldowndemo_run}
\title{Runs an example Shiny app}
\usage{
drilldowndemo_run(app_name = NULL)
}
\arguments{
\item{app_name}{Name of one of Shiny apps within the package. Possible values:
intro, tabs, navigation, details}
}
\description{
Runs an example Shiny app
}
